Refactored out API request into useIssues custom hook
